[refactor] re-organize socket.io and api.ai codes into subApp
