+++ server/controllers/api/chatbot.js
@@ -1,24 +1,31 @@
-const express     = require('express');
-const app         = express();
-const http        = require('http').Server(app);
-const io          = require('socket.io')(http);
+const apiai       = require('apiai');
+const botsFamily  = {
+  happy : apiai(process.env.HAPPY_BOT),
+  thirsty: apiai(process.env.THIRSTY_BOT),
+  drowning: apiai(process.env.DROWN_BOT),
+  dark: apiai(process.env.DARK_BOT),
+  burnt: apiai(process.env.BURNT_BOT)
+};
 
-io.on('connection', function(socket){
-  console.log('a user connected');
-  socket.emit('login','Hello, beautiful!'); //on client side, socket.on('login', (msg)=>{ $('<li>'').text(msg) })
-  socket.on('chat message', function(msg){
-    if (msg ==='Hello!') {
-      io.emit('chat message', 'What a wonderful day, Phoebe!');
-    } else if (msg ==='Who are you?') {
-      io.emit('chat message', 'I\'m Eric!');
-    } else if (msg === 'What are you doing?') {
-      io.emit('chat message', 'In a meeting with Benji');
-    } else if (msg === 'Are you a robot?') {
-      io.emit('chat message', 'No, I am Eric!');
-    } else {
-      io.emit('chat message', 'I don\'t understand you');
-    }
-    
+
+let plantbot;
+module.exports=function (io) {
+  io.on('connection', function(socket){
+    //once socket.io is connected 
+    //make an api call to fetch plant status
+    //diverse plantbot to handle client's questions based on plant status (i.e. happy, thirsty, drowning, burnt, dark)
+    plantbot = botsFamily.happy;//now it's hard coded to happy
+    io.emit('login','Hello');
+    socket.on('client', function(msg){
+      const request = plantbot.textRequest(msg);
+      request.on('response', function(response) {
+        //console.log(response.result.fulfillment.speech,'response from happybot');
+        io.emit('plant',response.result.fulfillment.speech);
+      });
+      request.on('error', function(error) {
+          console.log(error);
+      });
+      request.end();        
+    });
   });
-
-});
+};
